s/CurrentYear/ReplaceCurrentYearPlaceholder
